client/src/main/java/com/john/lewis/partnership/client/entity/deserializer/PriceNowDeserializer.java
package com.john.lewis.partnership.client.entity.deserializer;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.core.ObjectCodec;
import com.fasterxml.jackson.databind.*;

import java.io.IOException;
import java.math.BigDecimal;

public class PriceNowDeserializer extends JsonDeserializer<BigDecimal> {
    @Override
    public BigDecimal deserialize(JsonParser jsonParser, DeserializationContext ctxt) throws IOException {
        JsonNode node = jsonParser.getCodec().readTree(jsonParser);

        if(node.isObject()) {
            return new BigDecimal(node.get("from").asText());
        }

        return new BigDecimal(node.asText());
    }
}
